login logic code refactor
